Split Fetchers::Base's three engine concerns into modules
base.rb had grown to 944 lines doing three jobs the doc already describes
separately, so each becomes a mixin:

  CandidateWalk       where a gadget could start
  Disassembly         reading the file and finding its terminal calls
  ArgumentResolution  what a reached call requires of its caller

Every architecture hook stays declared in Base, whichever module holds the
engine that calls it: the class is the page a new architecture is written
against, and spreading that contract over three files would cost more than the
lines are worth. Base keeps them, the lifecycle, the string/global helpers the
arches build on, and the constraint folding, at 345 lines.

The corpus is byte-identical at levels 0/1/2 across every fixture.
